implement escape
