+++ webdoc/markup/markup.rb
@@ -1,0 +1,6 @@
+
+module Markup
+  def self.parse(str)
+    str.gsub('<', '&lt;')
+  end
+end
